requirements.txt
anyascii==0.3.3
asgiref==3.11.0
beautifulsoup4==4.14.2
bleach==4.1.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.2.8
django-filter==25.2
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.2.7
django-taggit==6.1.0
django-tasks==0.9.0
django-treebeard==4.7.1
djangorestframework==3.16.1
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
fonttools==4.61.0
idna==3.11
laces==0.1.2
Markdown==3.10
modelsearch==1.1.1
openpyxl==3.1.5
packaging==25.0
pillow==12.0.0
pillow_heif==1.1.1
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.12.1
pyphen==0.17.2
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
telepath==0.3.1
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wagtail==7.2.1
wagtail-markdown==0.13.0
wagtail-metadata==5.0.0
wagtail-resume==2.10.2
weasyprint==65.1
webencodings==0.5.1
Willow==1.12.0
zopfli==0.4.0
